Add doctree redirects (#14)
This should add the ability to redirect the pickle'd sources sphinx generates.  I'm hitting some gradle caching issues so want to be able to move them out of the gradle task output folder and into a different directory.  

Sphinx docs on the flag is here: 
https://www.sphinx-doc.org/en/master/man/sphinx-build.html#cmdoption-sphinx-build-d

Added source and testing for the new option.  

Do you need me to do anything else for it?  Is the wording okay?  I tried to stay as close to your style as possible. :)